requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
colorama==0.4.6
colorlog==6.8.2
cryptography==42.0.5
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-redis==5.4.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
filelock==3.13.4
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
identify==2.5.35
idna==3.7
inflection==0.5.1
itsdangerous==2.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
msgpack==1.0.8
multidict==6.0.5
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
pre-commit==3.7.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
ruff==0.3.5
six==1.16.0
sqlparse==0.4.4
twilio==9.0.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
whitenoise==6.6.0
yarl==1.9.4
